240118 09:45 김대호 Balance를 Double 에서 Int로 변경

+++ JavaMiniProject/src/controller/BalanceController.java
@@ -4,10 +4,10 @@ import model.UserDAO;
 import model.UserVO;
 
 public class BalanceController {
-	public double getBalance(UserVO member) {
+	public int getBalance(UserVO member) {
 		
 		UserDAO dao = new UserDAO();
-		double bal = dao.Balance(member);
+		int bal = dao.Balance(member);
 		return bal;
